:sparkles: Implementa Flyway migrations.

+++ src/main/java/com/example/authspringsecurityapp/domain/product/ProductRequestDTO.java
@@ -3,10 +3,12 @@ package com.example.authspringsecurityapp.domain.product;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProductRequestDTO(
         @NotBlank
         String name,
         @NotNull
-        Integer price
+        BigDecimal price
 ) {
 }
