+++ src/main/java/com/example/authspringsecurityapp/domain/product/ProductResponseDTO.java
@@ -1,6 +1,8 @@
 package com.example.authspringsecurityapp.domain.product;
 
-public record ProductResponseDTO(String id, String name, Integer preco) {
+import java.math.BigDecimal;
+
+public record ProductResponseDTO(String id, String name, BigDecimal preco) {
     public ProductResponseDTO(Product product) {
         this(product.getId(), product.getName(), product.getPrice());
     }
